Refactor chasm/lib/nexusoperation/workflow into chasm/lib/workflow
- Moved Nexus operation event definitions, command handlers, and library
  registration from `chasm/lib/nexusoperation/workflow` into
  `chasm/lib/workflow` as `nexus_events.go`, `nexus_commands.go`,
  `nexus_library.go`.
- Extracted all Nexus-related Workflow methods from `workflow.go` into
  `nexus_methods.go`.
- Converted all `AddAndApplyHistoryEventByEventType` calls to use the
  generic `addAndApplyHistoryEvent[D]` form and removed
  `AddAndApplyHistoryEventByEventType` entirely.
- Unexported internal symbols: `addAndApplyHistoryEvent`,
  `eventDefinitionByGoType`, `addNexusOperation`,
  `removeNexusOperation`, `pendingNexusOperationCount`.
- Absorbed the `chasmnexusworkflow.Module` fx registration into the
  workflow module's `fx.Invoke`, removing the separate fx module.
- Deleted `chasm/lib/nexusoperation/workflow/` package.

The `chasm/lib/workflow` package is a more natural home for this code.
It allows all of the Nexus integration for workflows to be colocated
alongside the workflow component, registry, and event application logic
rather than spread across a separate package that existed solely to
register handlers with the workflow registry.

- `go build ./...` passes.
- `go test ./chasm/lib/workflow/... -count 1` passes (all existing tests
  moved and adapted).
- `make lint` passes with 0 issues.
